FEATURE: Ingress Caddy Proxy as separate minimal Docker Compose project
This creates a separate, minimal "ingress" docker-compose setup with Caddy.

The main benefit is that this way, we can
easily host multiple projects for the same
customer on a single server.

It does:
- TLS termination
- domain redirects
- reverse proxying to the application(s)
- /up-ingress health check URL
- (potentially) display of error pages if site down
- (potentially) access logs to NATS

It is an extension point, because you can use
any Caddy Module (or quite easily write your
own in Go).