fix: skipping a card won't crash the game
Skipping a card could lead to a game crash when only one Zeitsteinslot is available in a category. The skip action would consume a Zeitsteinslot, leaving no free slot. The next Card could then not be activated, because no free slot was available. After a SkipCard action an ActivateCard action should not need a Zeitsteinslot, though.

This commit allows an ActivateCard action with no free Zeitsteinslots **after** a SkipCard action.

Closes: #312