fix: only update gameEvents when necessary
Calling `ForCoreGameLogic::getGameEvents()` all the time was having a really heavy performance impact. Now we use just save it to the GameUi and update it everytime we call a command.
To make sure that we don't forget to use the `GameUi::handle()` method, we added a phpstan rule that disallows calls to `ForCoreGameLogic::getGameEvents()` inside of traits.