task: use a dedicated branch for staging
WHY: We want to allow the team from the University to test the game
while we still work on it. This way we can continue to develop against
the `main` branch and deploy hotfixes to `staging` if needed.

We are now at a stage of development, where we want to be more delicate
about pushing changes to staging.